refactor: move StatementType and StatementFrequency from common to financials
These enums are only used by the financials endpoint, so they belong in the financials module rather than common. This keeps common.rs focused on truly shared types like CandleResolution that are used across multiple asset types.
